requirements.txt
pandas==1.4.2
python-dotenv==0.20.0
scikit_learn==1.0.2
spotipy==2.19.0
